requirements.txt

Django==4.1.1
numpy==1.23.4
Pillow==9.2.0
pytesseract==0.3.10